[BLE] move demo_peripheral.rb to example/peripheral/app.rb
